# backend/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChronoSift API",
    "description": "Forensic Timeline Analysis API - Comprehensive documentation for all endpoints",
    "version": "1.0.0",
    "contact": {
      "name": "ChronoSift",
      "url": "https://github.com/yourusername/chronosift"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      }
    },
    "schemas": {
...
      "ReportTemplate": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "description": {"type": "string"},
          "project_id": {"type": "integer"},
          "template_content": {"type": "string"},
          "config": {"type": "object"},
          "is_public": {"type": "boolean"},
          "category": {"type": "string"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      }
    }
  },
  "paths": {
...
    "/reports/templates": {
      "post": {
        "tags": ["Reports"],
        "summary": "Create a new report template",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "template_content", "project_id"],
                "properties": {
                  "name": {"type": "string"},
                  "description": {"type": "string"},
                  "project_id": {"type": "integer"},
                  "template_content": {"type": "string", "description": "HTML/Jinja2 template"},
                  "config": {
                    "type": "object",
                    "properties": {
                      "page_size": {"type": "string", "enum": ["A4", "Letter", "Legal"]},
                      "orientation": {"type": "string", "enum": ["portrait", "landscape"]}
                    }
                  },
                  "is_public": {"type": "boolean"},
                  "category": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ReportTemplate"}
              }
            }
          }
        }
      }
    },
    "/reports/projects/{project_id}/templates": {
      "get": {
        "tags": ["Reports"],
        "summary": "List report templates for a project",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "List of templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/ReportTemplate"}
                }
              }
            }
          }
        }
      }
    },
    "/reports/generate": {
      "post": {
        "tags": ["Reports"],
        "summary": "Generate a report from a template",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["template_id", "project_id"],
                "properties": {
                  "template_id": {"type": "integer"},
                  "project_id": {"type": "integer"},
                  "timeline_id": {"type": "integer", "nullable": true},
                  "start_date": {"type": "string", "format": "date-time"},
                  "end_date": {"type": "string", "format": "date-time"},
                  "entry_limit": {"type": "integer", "default": 1000},
                  "name": {"type": "string"},
                  "description": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "integer"},
                    "name": {"type": "string"},
                    "pdf_path": {"type": "string"},
                    "entry_count": {"type": "integer"},
                    "file_size": {"type": "integer"},
                    "generation_time": {"type": "number"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reports/reports/{report_id}/download": {
      "get": {
        "tags": ["Reports"],
        "summary": "Download a generated report as PDF",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  }
}